add functionality to hera application
