+++ lms/djangoapps/hera/urls.py
@@ -7,5 +7,6 @@ from django.conf.urls import url
 from hera import views
 
 urlpatterns = [
-    url(r'^onboarding/', views.OnboardingPagesView.as_view(), name='onboarding')
+    url(r'^onboarding/', views.OnboardingPagesView.as_view(), name='onboarding'),
+    url(r'^selection_page/', views.SelectionPageView.as_view(), name='selection_page'),
 ]
