--- lms/djangoapps/selection_page/apps.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
-from django.apps import AppConfig
-
-
-class SelectionPageConfig(AppConfig):
-    name = 'selection_page'
